backend/requirements.txt
flask
python-magic
yara-python
pefile
flask-cors
python-docx
PyPDF2